refactor: PersistentDataTypeRegistry
