+++ server/src/main/java/org/allaymc/server/pdc/AllayPersistentDataContainer.java
@@ -2,6 +2,7 @@ package org.allaymc.server.pdc;
 
 import org.allaymc.api.pdc.PersistentDataContainer;
 import org.allaymc.api.pdc.PersistentDataType;
+import org.allaymc.api.pdc.PersistentDataTypeRegistry;
 import org.allaymc.api.utils.Identifier;
 import org.cloudburstmc.nbt.NbtMap;
 import org.cloudburstmc.nbt.NbtUtils;
